requirements.txt
# Core libraries
numpy
pandas
matplotlib

# Machine Learning
tensorflow  # Or the latest stable version
scikit-learn

# (Optional) Web App
flask             # If you're using Flask
streamlit         # If you're using Streamlit (choose one)
flask-cors
# Image handling
Pillow
requests       # For working with uploaded images
